refactor: split swiftdatabookrepo migration into extension file
